Updated with tests
