+++ .github/workflows/publish-beta.yml
@@ -1,0 +1,35 @@
+name: Publish Beta Package
+
+on:
+  pull_request:
+    branches:
+      - 'main'
+
+env:
+  CI: true
+  NODE_OPTIONS: --experimental-vm-modules
+
+jobs:
+  comment:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node-version: [18.x]
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          ref: ${{ github.event.pull_request.head.sha }}
+      - name: Use Node.js ${{ matrix.node-version }}
+        uses: actions/setup-node@v3
+        with:
+          node-version: ${{ matrix.node-version }}
+          cache: 'npm'
+      - name: Install dependencies
+        run: npm ci
+      - name: Run Test
+        run: npm test
+      - name: perform beta publish
+        uses: checkdigit/github-actions/publish-beta@main
+        env:
+          GITHUB_TOKEN: ${{secrets.GITHUB_TOKEN}}
+          NPM_TOKEN: ${{ secrets.NPM_TOKEN_PUBLISH }}
