+++ package.json
@@ -8,7 +8,7 @@
   },
   "peerDependencies": {
     "@types/node": ">=16",
-    "typescript": ">=5.0.2 <5.1"
+    "typescript": ">=5.0.4 <5.1"
   },
   "repository": {
     "type": "git",
@@ -23,16 +23,50 @@
   "scripts": {
     "preversion": "npm test",
     "postversion": "git push && git push --tags",
-    "prettier": "prettier --list-different .",
-    "prettier:fix": "prettier --write .",
+    "prettier": "prettier --ignore-path .gitignore --list-different .",
+    "prettier:fix": "prettier --ignore-path .gitignore --write .",
     "test": "npm run ci:compile && npm run ci:test && npm run ci:style",
-    "ci:compile": "tsc --noEmit",
-    "ci:test": "tsc && node build/index.js | grep -q 'complete' && rimraf build",
+    "build-tsc": "rimraf build && tsc",
+    "build-es": "rimraf build-es && esbuild ./src/* --platform=node --bundle --format=esm --sourcemap=inline --outdir=build-es",
+    "build-swc": "rimraf build-swc && swc ./src -d ./build-swc",
+    "ci:test": "NODE_OPTIONS=\"--no-warnings --experimental-vm-modules\" jest --coverage=false",
+    "ci:compile": "npm run build-tsc && npm run build-es",
     "ci:style": "npm run prettier"
   },
   "devDependencies": {
-    "@checkdigit/prettier-config": "^3.3.0",
-    "rimraf": "^4.4.0"
+    "@checkdigit/prettier-config": "^3.4.0",
+    "@swc/cli": "^0.1.62",
+    "@swc/core": "^1.3.50",
+    "@types/jest": "^29.5.0",
+    "esbuild": "^0.17.16",
+    "get-port": "^6.1.2",
+    "got": "11.8.6",
+    "jest": "^29.5.0",
+    "rimraf": "^5.0.0",
+    "ts-jest": "^29.1.0"
+  },
+  "jest": {
+    "extensionsToTreatAsEsm": [
+      ".mts"
+    ],
+    "transform": {
+      "^.+\\.ts$": [
+        "ts-jest",
+        {
+          "isolatedModules": true,
+          "diagnostics": false,
+          "useESM": false
+        }
+      ]
+    },
+    "collectCoverageFrom": [
+      "<rootDir>/src/**",
+      "!<rootDir>/src/**/*.spec.ts",
+      "!<rootDir>/src/**/*.test.ts"
+    ],
+    "testMatch": [
+      "<rootDir>/src/**/*.spec.ts"
+    ]
   },
   "files": [
     "tsconfig.json",
