+++ src/module-directory/index.ts
@@ -1,0 +1,3 @@
+// module-directory/index.ts
+
+export default () => 'module-directory-index';
