+++ tsconfig.json
@@ -1,8 +1,7 @@
 {
   "compilerOptions": {
     "module": "commonjs",
-    "target": "es2022",
-    "lib": ["esnext", "dom", "webworker"],
+    "target": "esnext",
     "sourceMap": true,
     "inlineSources": true,
     "outDir": "build",
